implement matrix decomposition recommender using Vowpal Wabbit
Achieves an R^2 of about 0.165 on the test set using a random 50/50 split. On subsequent random splits (which include training data), gets about 0.23. So might be overfitting a little bit. Will test it later on a smaller dataset limited to just restaurants in an area, because that's more relevant for meaningful predictions.